pyproject.toml
[tool.poetry]
name = "pollster"
version = "0.1.0"
description = "Polls app on Django"
authors = ["mauer9 <75573495+mauer9@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
django = "~4.2.7"
django-stubs = {extras = ["compatible-mypy"], version = "~4.2.6"}
faker = "~20.1.0"
psycopg2-binary = "~2.9.9"
gunicorn = "~21.2.0"
whitenoise = "~6.6.0"
python-dotenv = "~1.0.0"
django-debug-toolbar = "^4.3.0"
django-redis = "^5.4.0"
icecream = "^2.1.3"
pika = "^1.3.2"
celery = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
